+++ src/MapView/MapView/index.ts
@@ -1,0 +1,4 @@
+import MapView from './MapView';
+
+export type {MapViewProps, MapViewHandle} from './MapViewTypes';
+export default MapView;
